// apps/frontend/src/generated/queries.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  JSON: any;
};
// ...
export type Course = {
  __typename?: 'Course';
  /** course creation date. */
  createdAt: Scalars['DateTime'];
  /** course description in markup. */
  description: Scalars['String'];
  /** amount of discount that is applied to course price. */
  discountPercent: Scalars['String'];
  /** total hours of course. */
  hoursCount: Scalars['Int'];
  /** course id. */
  id: Scalars['ID'];
  /** resolve image field. */
  image: CourseImage;
  instructors: Array<Instructor>;
  /** total lectures of course. */
  lecturesCount: Scalars['Int'];
  /** course level which is an enum. */
  level: CourseLevel;
  /** total participants of course. */
  participantsCount: Scalars['Int'];
  /** description for pre requirements in markup */
  preRequirementsDescription?: Maybe<Scalars['String']>;
  /** price of course. */
  price: Scalars['Float'];
  /** progress that made for course. */
  progress: Scalars['Float'];
  /** short description that is shown in courses list. */
  shortDescription: Scalars['String'];
  /** spot player created course code. */
  spotPlayerCourseId: Scalars['String'];
  /** course title. */
  title: Scalars['String'];
  /** course update date. */
  updatedAt: Scalars['DateTime'];
};

export type CourseImage = {
  __typename?: 'CourseImage';
  /** related course. */
  course: Course;
  /** course cover image that is only shown in single course pages. */
  coverImage: Scalars['String'];
  id: Scalars['ID'];
  /** course image, image which is course main image. */
  image: Scalars['String'];
};

export enum CourseLevel {
  Advanced = 'ADVANCED',
  Elementary = 'ELEMENTARY',
  Intermediate = 'INTERMEDIATE',
  Mixed = 'MIXED'
}
// ...
export type Expertise = {
  __typename?: 'Expertise';
  /** expertise id. */
  id: Scalars['ID'];
  /** instructor that expertise relates to. */
  instructor: Instructor;
  /** determine wether an expertise is instructor's primary or not. */
  isPrimary: Scalars['Boolean'];
  /** expertise label/title. */
  label: Scalars['String'];
  /** expertise level which is an enum. */
  level: ExpertiseLevel;
  /** determine wether expertise is validated by admin or not. */
  validated: Scalars['Boolean'];
};
// ...
export type Instructor = {
  __typename?: 'Instructor';
  /** a short description about instructor. */
  about: Scalars['String'];
  /** courses that instructor is the instructor of them. */
  courses?: Maybe<Array<Course>>;
  /** instructor creation date. */
  createdAt: Scalars['DateTime'];
  expertises: Array<Expertise>;
  /** instructor id. */
  id: Scalars['ID'];
  /** list of instructor social medias. */
  socials?: Maybe<Array<Expertise>>;
  /** instructor update date. */
  updatedAt: Scalars['DateTime'];
  /** returns logged in user credentials. */
  user: User;
};
// ...
export type User = {
  __typename?: 'User';
  /** all of user answered question. */
  answeredQuestions?: Maybe<Array<Question>>;
  /** all of user asked questions. */
  askedQuestions?: Maybe<Array<Question>>;
  avatar?: Maybe<Avatar>;
  cart: Cart;
  /** all user comments. */
  comments?: Maybe<Array<Comment>>;
  /** all of user bought courses. */
  courses?: Maybe<Array<Course>>;
  /** user creation date. */
  createdAt: Scalars['DateTime'];
  /** user created category, its useful when user is admin or instructor. */
  createdCategories?: Maybe<Array<Category>>;
  /** user email which is unique and must for sign up. */
  email: Scalars['String'];
  /** user firstname. */
  firstname?: Maybe<Scalars['String']>;
  /** user id. */
  id: Scalars['ID'];
  /** if user is an instructor this field relates to. */
  instructor?: Maybe<Instructor>;
  /** user lastname. */
  lastname?: Maybe<Scalars['String']>;
  /** all of user orders. */
  orders?: Maybe<Array<Order>>;
  /** user phone number which is unique. */
  phoneNumber?: Maybe<Scalars['String']>;
  /** user role which is enum. */
  role: UserRole;
  /** user score gain rate which depends on user activity. */
  scoreGainRate: Scalars['Float'];
  /** total scores gained by user. */
  scores: Scalars['Float'];
  /** user update date. */
  updatedAt: Scalars['DateTime'];
};